refactor: split cli option parsing
