+++ tests/run.sh
@@ -22,14 +22,15 @@ cc -Iinclude -Wall -Wextra -std=c99 \
     src/ast_expr.c src/ast_stmt_create.c src/ast_stmt_free.c src/lexer.c src/util.c \
     src/vector.c src/error.c
 # build cli unit test binary with vector_push wrapper
-cc -Iinclude -Wall -Wextra -std=c99 -Dvector_push=test_vector_push -c src/cli.c -o cli_test.o
-cc -Iinclude -Wall -Wextra -std=c99 -c src/cli_env.c -o cli_env_test.o
+cc -Iinclude -Wall -Wextra -std=c99 -Dvector_push=test_vector_push -c src/cli_opts_parse.c -o cli_parse_test.o
+cc -Iinclude -Wall -Wextra -std=c99 -c src/cli_opts_env.c -o cli_env_test.o
 cc -Iinclude -Wall -Wextra -std=c99 -Dvector_push=test_vector_push -c src/cli_opts.c -o cli_opts_test.o
 cc -Iinclude -Wall -Wextra -std=c99 -c "$DIR/unit/test_cli.c" -o "$DIR/test_cli.o"
+cc -Iinclude -Wall -Wextra -std=c99 -c "$DIR/unit/test_cli_opts_parse.c" -o "$DIR/test_cli_opts_parse.o"
 cc -Iinclude -Wall -Wextra -std=c99 -c src/vector.c -o vector_test.o
 cc -Iinclude -Wall -Wextra -std=c99 -DUNIT_TESTING -DNO_VECTOR_FREE_STUB -c src/util.c -o util_test.o
-cc -o "$DIR/cli_tests" cli_test.o cli_env_test.o cli_opts_test.o "$DIR/test_cli.o" vector_test.o util_test.o
-rm -f cli_test.o cli_env_test.o cli_opts_test.o "$DIR/test_cli.o" vector_test.o util_test.o
+cc -o "$DIR/cli_tests" cli_parse_test.o cli_env_test.o cli_opts_test.o "$DIR/test_cli.o" "$DIR/test_cli_opts_parse.o" vector_test.o util_test.o
+rm -f cli_parse_test.o cli_env_test.o cli_opts_test.o "$DIR/test_cli.o" "$DIR/test_cli_opts_parse.o" vector_test.o util_test.o
 # build parser alloc failure unit test with vector_push wrapper
 cc -Iinclude -Wall -Wextra -std=c99 -Dvector_push=test_vector_push -c src/parser_core.c -o parser_core_fail.o
 cc -Iinclude -Wall -Wextra -std=c99 -c src/parser_init.c -o parser_init_fail.o
